# docs/swagger.yaml
basePath: /
definitions:
...
  request.RondaScheduleCreateRequest:
    properties:
      efektif_date:
        type: string
      group_id:
        type: string
    type: object
  request.RondaScheduleUpdateRequset:
    properties:
      efektif_date:
        type: string
      group_id:
        type: string
      id:
        type: string
    type: object
  request.ShopCreateRequest:
    properties:
      description:
        type: string
      name:
        type: string
      status:
        description: aktif/nonaktif
        type: string
      user_id:
        type: integer
    type: object
...
  request.ShopUpdateRequset:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      status:
        description: aktif/nonaktif
        type: string
      user_id:
        type: integer
    type: object
...
  utils.Pagination:
    properties:
      limit:
        type: integer
      page:
        type: integer
      sort:
        type: string
      sort_by:
        type: string
      total_pages:
        type: integer
      total_rows:
        type: integer
    type: object
  utils.ResponseData:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Pagination'
      status:
        type: string
    type: object
host: localhost:5001
info:
  contact:
    email: sarikhin@yahoo.co.id
    name: API Support
  description: This is a API Server for SIRNAWA
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SIRNAWA API
  version: "1.0"
paths:
...
  /ronda-schedule:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get List RondaSchedule
      tags:
      - RondaSchedule
    post:
      consumes:
      - application/json
      parameters:
      - description: Create RondaSchedule
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RondaScheduleCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Create RondaSchedule
      tags:
      - RondaSchedule
  /ronda-schedule/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: RondaSchedule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Delete RondaSchedule By ID
      tags:
      - RondaSchedule
    get:
      consumes:
      - application/json
      parameters:
      - description: RondaSchedule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Find RondaSchedule By ID
      tags:
      - RondaSchedule
    put:
      consumes:
      - application/json
      parameters:
      - description: Update RondaSchedule
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.RondaScheduleUpdateRequset'
      - description: RondaSchedule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Update RondaSchedule
      tags:
      - RondaSchedule
  /ronda-schedule/paginated:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Order by
        in: query
        name: order_by
        type: string
      - description: Order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get Paginated RondaSchedule
      tags:
      - RondaSchedule
  /shop:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get List Shop
      tags:
      - Shop
    post:
      consumes:
      - application/json
      parameters:
      - description: Create Shop
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ShopCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Create Shop
      tags:
      - Shop
...
  /shop/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Shop id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Delete Shop By ID
      tags:
      - Shop
    get:
      consumes:
      - application/json
      parameters:
      - description: Shop id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Find Shop By ID
      tags:
      - Shop
    put:
      consumes:
      - application/json
      parameters:
      - description: Update Shop
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ShopUpdateRequset'
      - description: Shop id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Update Shop
      tags:
      - Shop
  /shop/paginated:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Order by
        in: query
        name: order_by
        type: string
      - description: Order
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ResponseData'
      security:
      - Bearer: []
      summary: Get Paginated Shop
      tags:
      - Shop
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"